# resources/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WuKongIM API",
    "description": "WuKongIM is a high-performance instant messaging system. This API documentation covers all REST endpoints for managing users, channels, messages, connections, and system operations.",
    "version": "2.0.0",
    "contact": {
      "name": "WuKongIM Team",
      "url": "https://github.com/WuKongIM/WuKongIM"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5001",
      "description": "Local development server"
    },
    {
      "url": "https://api.wukongim.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/route": {
      "get": {
        "summary": "Get User IM Address",
        "description": "Get the IM connection address for users",
        "tags": ["Route"],
        "parameters": [
          {
            "name": "intranet",
            "in": "query",
            "description": "Whether to return intranet address (1 for true, 0 for false)",
            "schema": {
              "type": "integer",
              "enum": [0, 1],
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "IM connection addresses",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IMAddressResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/channel": {
      "post": {
        "summary": "Create or Update Channel | 创建或更新频道",
        "description": "Create a new channel or update an existing channel with basic information and initial subscribers. This is the primary endpoint for channel management.\n\n**功能说明：**\n- 创建新频道或更新现有频道\n- 设置频道基本信息（大群、禁言、解散等状态）\n- 添加初始订阅者（不支持个人频道）\n- 支持重置订阅者列表\n- 个人频道支持陌生人消息设置\n\n**Channel Types:**\n- 1: Person channel (个人频道) - Direct messaging\n- 2: Group channel (群组频道) - Group messaging\n- Other values: Custom channel types\n\n**Important Notes:**\n- Person channels (channel_type=1) do not support subscribers\n- Use reset=1 to replace all existing subscribers\n- Special characters in channel_id are not allowed\n- System will automatically route to appropriate cluster node",
        "tags": ["Channel"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChannelCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Channel created/updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missing_channel_id": {
                    "summary": "Missing channel ID",
                    "value": {
                      "error": "channel_id不能为空！"
                    }
                  },
                  "invalid_channel_type": {
                    "summary": "Invalid channel type",
                    "value": {
                      "error": "频道类型不能为0！"
                    }
                  },
                  "special_characters": {
                    "summary": "Special characters in channel ID",
                    "value": {
                      "error": "频道ID不能包含特殊字符！"
                    }
                  },
                  "person_channel_subscribers": {
                    "summary": "Person channel with subscribers",
                    "value": {
                      "error": "个人频道不支持订阅者！"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "cluster_error": {
                    "summary": "Cluster node error",
                    "value": {
                      "error": "获取频道所在节点失败！"
                    }
                  },
                  "create_channel_error": {
                    "summary": "Create channel error",
                    "value": {
                      "error": "创建频道失败！"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/message": {
      "post": {
        "summary": "Search Single Message",
        "description": "Search for a single message",
        "tags": ["Message"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageSearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/manager/login": {
      "post": {
        "summary": "Manager Login",
        "description": "Login for manager access",
        "tags": ["Manager"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManagerLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagerLoginResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "IMAddressResponse": {
        "type": "object",
        "properties": {
          "tcp_addr": {
            "type": "string",
            "example": "127.0.0.1:5100"
          },
          "ws_addr": {
            "type": "string",
            "example": "ws://127.0.0.1:5200"
          },
          "wss_addr": {
            "type": "string",
            "example": "wss://127.0.0.1:5300"
          }
        }
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "ok"
          }
        }
      },
...
      "ChannelCreateRequest": {
        "type": "object",
        "required": ["channel_id", "channel_type"],
        "description": "Request to create or update a channel with its basic information and initial subscribers.",
        "properties": {
          "channel_id": {
            "type": "string",
            "description": "Channel ID. Cannot be empty or contain special characters.",
            "example": "group123",
            "minLength": 1,
            "maxLength": 64
          },
          "channel_type": {
            "type": "integer",
            "format": "uint8",
            "description": "Channel type. Cannot be 0. (1=person, 2=group, etc.)",
            "example": 2,
            "minimum": 1,
            "maximum": 255
          },
          "large": {
            "type": "integer",
            "enum": [0, 1],
            "example": 0,
            "description": "Whether it's a large group (0=no, 1=yes). Large groups have different performance characteristics."
          },
          "ban": {
            "type": "integer",
            "enum": [0, 1],
            "example": 0,
            "description": "Whether the channel is banned (0=no, 1=yes). Banned channels prevent all users from sending/receiving messages except system accounts."
          },
          "disband": {
            "type": "integer",
            "enum": [0, 1],
            "example": 0,
            "description": "Whether the channel is disbanded (0=no, 1=yes). Disbanded channels are marked as deleted."
          },
          "send_ban": {
            "type": "integer",
            "enum": [0, 1],
            "example": 0,
            "description": "Whether sending messages is banned (0=no, 1=yes). When enabled, all members cannot send messages but can receive them. For person channels, the user can receive but not send messages."
          },
          "allow_stranger": {
            "type": "integer",
            "enum": [0, 1],
            "example": 0,
            "description": "Whether to allow strangers to send messages (0=no, 1=yes). Currently only supported for person channels. When enabled, strangers can send messages to the user."
          },
          "reset": {
            "type": "integer",
            "enum": [0, 1],
            "example": 0,
            "description": "Whether to reset subscribers (0=no, 1=yes). When enabled, all existing subscribers will be removed before adding new ones."
          },
          "subscribers": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of initial subscriber user IDs. Not supported for person channels (channel_type=1).",
            "example": ["user1", "user2", "user3"]
          }
        },
        "examples": {
          "create_group": {
            "summary": "Create a group channel",
            "value": {
              "channel_id": "group123",
              "channel_type": 2,
              "large": 0,
              "ban": 0,
              "send_ban": 0,
              "subscribers": ["user1", "user2", "user3"]
            }
          },
          "create_person_channel": {
            "summary": "Create a person channel",
            "value": {
              "channel_id": "user123",
              "channel_type": 1,
              "allow_stranger": 1,
              "send_ban": 0
            }
          },
          "reset_subscribers": {
            "summary": "Reset channel subscribers",
            "value": {
              "channel_id": "group123",
              "channel_type": 2,
              "reset": 1,
              "subscribers": ["new_user1", "new_user2"]
            }
          }
        }
      },
...
      "MessageHeader": {
        "type": "object",
        "properties": {
          "no_persist": {
            "type": "integer",
            "enum": [0, 1],
            "example": 0,
            "description": "Whether message should not be persisted (0=persist, 1=no persist)"
          },
          "red_dot": {
            "type": "integer",
            "enum": [0, 1],
            "example": 1,
            "description": "Whether to show red dot notification (0=no, 1=yes)"
          },
          "sync_once": {
            "type": "integer",
            "enum": [0, 1],
            "example": 0,
            "description": "Whether message should be synced only once (0=no, 1=yes)"
          }
        }
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "header": {
            "$ref": "#/components/schemas/MessageHeader"
          },
          "setting": {
            "type": "integer",
            "format": "uint8",
            "description": "Message setting flags.",
            "example": 0
          },
          "message_id": {
            "type": "integer",
            "format": "int64",
            "example": 123456789
          },
          "message_idstr": {
            "type": "string",
            "description": "String representation of message_id.",
            "example": "123456789"
          },
          "message_seq": {
            "type": "integer",
            "format": "uint64",
            "example": 1001
          },
          "client_msg_no": {
            "type": "string",
            "example": "client_msg_123"
          },
          "event_meta": {
            "$ref": "#/components/schemas/MessageEventMeta"
          },
          "from_uid": {
            "type": "string",
            "example": "user123"
          },
          "channel_id": {
            "type": "string",
            "example": "group123"
          },
          "channel_type": {
            "type": "integer",
            "format": "uint8",
            "example": 2
          },
          "topic": {
            "type": "string",
            "example": "topic_1"
          },
          "expire": {
            "type": "integer",
            "format": "uint32",
            "example": 0
          },
          "timestamp": {
            "type": "integer",
            "format": "int32",
            "example": 1640995200
          },
          "payload": {
            "type": "string",
            "format": "byte",
            "example": "SGVsbG8gV29ybGQ="
          }
        }
      },
      "MessageEventMeta": {
        "type": "object",
        "properties": {
          "has_events": {
            "type": "boolean",
            "description": "Whether this message has associated event states.",
            "example": true
          },
          "event_version": {
            "type": "integer",
            "format": "uint64",
            "description": "Version of event projection (currently equals last_msg_event_seq).",
            "example": 12
          },
          "last_msg_event_seq": {
            "type": "integer",
            "format": "uint64",
            "description": "Largest msg_event_seq among all event keys.",
            "example": 12
          },
          "event_key_count": {
            "type": "integer",
            "description": "Count of event keys associated with this message.",
            "example": 2
          },
          "open_event_key_count": {
            "type": "integer",
            "description": "Count of event keys still in open state.",
            "example": 1
          },
          "event_keys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageEventKeyMeta"
            }
          }
        }
      },
      "MessageEventKeyMeta": {
        "type": "object",
        "properties": {
          "event_key": {
            "type": "string",
            "description": "Event key identifier inside the message.",
            "example": "main"
          },
          "status": {
            "type": "string",
            "description": "Current status of this event key projection.",
            "enum": ["open", "closed", "error", "cancelled"],
            "example": "open"
          },
          "last_msg_event_seq": {
            "type": "integer",
            "format": "uint64",
            "description": "Last msg_event_seq applied to this event key.",
            "example": 12
          },
          "snapshot": {
            "description": "Optional merged snapshot payload (present in full summary mode).",
            "nullable": true
          },
          "end_reason": {
            "type": "integer",
            "format": "uint8",
            "description": "Terminal reason code for the typed event lane.",
            "example": 0
          },
          "error": {
            "type": "string",
            "description": "Public terminal error code or text when present.",
            "example": ""
          }
        }
      },
...
      "MessageSearchRequest": {
        "type": "object",
        "required": ["message_id"],
        "properties": {
          "message_id": {
            "type": "integer",
            "format": "int64",
            "example": 123456789
          }
        }
      },
...
      "ManagerLoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": {
            "type": "string",
            "description": "Manager username",
            "example": "admin"
          },
          "password": {
            "type": "string",
            "description": "Manager password",
            "example": "password123"
          }
        }
      },
      "ManagerLoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Authentication token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "expires_in": {
            "type": "integer",
            "description": "Token expiration time in seconds",
            "example": 3600
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message",
            "example": "Invalid request"
          },
          "message": {
            "type": "string",
            "description": "Detailed error description",
            "example": "The provided parameters are invalid"
          }
        }
      }
    }
  },
...
}